[feature] adding components and style
