+++ src/App.css
@@ -1,4 +1,24 @@
 .appContainer {
   height: 100vh;
   width: 100vw;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  background-color: #18191a;
+}
+.appCard{
+  min-height: 500px;
+  min-width: 300px;
+  box-shadow: rgb(9 30 66 / 25%) 0px 1px 1px, rgb(9 30 66 / 13%) 0px 0px 1px 1px;
+  padding: 8px;
+  background-color: #fff;
+  border-radius: 12px;
+  display: grid;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  overflow-y: scroll;
+  overflow-x: hidden;
+  text-align: center;
 }
