+++ src/Core/Car.js
@@ -1,5 +1,4 @@
 import {v4 as uuidv4} from 'uuid';
-import Driver from './Driver';
 
 export default class Car {
     constructor(data) {
